# [7.x] Enable dot notation to set nested visibilities for eloquent serialization