This allows dot notation for the `visible` and `hidden` directives so one could show/hide a field on a relation. This enables things like:

```
$collection = MyModel::with('relatives', 'relative')->get();

$collection->makeHidden('relatives.id');
$collection->makeVisible('relatives', 'relative', 'relative.email');

$collection->toArray();
```

This uses the `hidden`/`visible` properties on the model and sets them accordingly on the related one. Any other way of setting visibilities would also work including `$hidden = [relative.address]` in the model class and `$model->setVisible(['friends.home.phone'])`.

Currently it is not possible to affect visibility on relations without explicitly setting them on the related models (or their collections) themselves.

A feature like this has been requested in #8945 as well as on [multiple](https://stackoverflow.com/questions/42390595/laravel-model-with-and-makehidden) [questions](https://stackoverflow.com/questions/57027777/how-can-i-chain-makehidden-when-eager-loading-a-belongstomany-relationship) on [other](https://laracasts.com/discuss/channels/eloquent/nested-attributes-in-makehidden) [sites](https://laracasts.com/discuss/channels/eloquent/makehidden-for-relationship-columns).

The usual workaround is to select only the needed columns when eager loading:

```
$collection = MyModel::with(['relatives' => function($relative) {
    $relative->select('id', 'name', 'nickname');
}])->get();
```

However, that's not a solution if you want to hide the foreign key (the relation just gets lost) or make visible something that's not visible on the related model.