@taylorotwell 2019-10-21
Have you done any performance tests on this? We're not looping through every model, and then for every model we're looping through every visible / hidden array attribute and performing explode operations on it.