@tontonsb 2019-10-21
I agree, that doesn't sound effective at all. Especially because the same 30 explodes will be repeated over and over.

I will test the performance and try to come up with a more savvy solution over this week.