@tontonsb 2019-10-29
I have done some thinking and some experiments. I discovered that there is a significant performance hit on empty models, but no measurable impact on more complex cases where the explodes seem to constitute below 1% of the runtime.

The actual impact of the proposed approach seems to be below 3E-7 seconds per directive per model.

## Performance on empty models

Taking empty model with no data or relations. 500 models, 15 properties on each. The timings were done using PHP's `microtime(true)` for the serialization (`toJson()`) only, on a typical $5 VPS. These results were concluded from 20 runs of each the old and new serialization:

|               | Old      | New     |
|----------|---------|---------| 
| Median  | 0.0039 | 0.0056 |
| Average | 0.0044 | 0.0062 |
| St. dev.   | 0.0022 | 0.0014 |

The 7500 additional explodes increase the run time by around 40%. But in absolute units it's only 2 milliseconds. This also suggests that every property on every model increases the cost by around 2.4E-7.

By taking more models (5000) I was able to get standard deviation below 10%. By varying both model count and directive count I evaluated the cost of a propery to be between 2.4E-7 and 2.8E-7.

## Performance on real models

For the real-life case I took 2500 models. Each model had 20 properties - 13 serialized and 7 hidden as well as up to four relations, each having 3 properties - 2 visible and 1 hidden. The total number of directives that had to be exploded on each model is 25.

The results of each were around 3.35 seconds. The standard deviation - around 0.28 seconds. That's regardless of using the old or the new approach or changing the number of visibility directives! I was not able to measure any impact and that's as expected - the total impact of 62500 explodes should be around 0.015-0.018s using the above estimate and that's below noise levels.

## Alternative solution

Another solution that does not require exploding is using this syntax like this:

```
$collection->makeHidden([
    'email',
    'username',
    'relative' => [
        'id',
        'username',
    ],
])
```

That would simplify the current PR (the job of explodes is to prepare array like that) but require to rewrite some of the other methods. This also ahs the benefit that one could `makeVisible` something on the relation without actually turning on the whitelist mode on the initial model (we could have some arrays but no string values in `$visible`) and having to list everything that should be visible on it.

I can write up that solution, but I personally find the dot notation to be more convenient and the performance hit doesn't appear noticable.