@donnysim 2019-10-31
We had cases where just casting model attributes cost +/- 50% of the request time for big lists, so I'd appreaciate if this didn't add another "needs a workaround" solution 😅