@taylorotwell 2019-10-31
@tontonsb how many other methods would have to be touched if we did the array structure syntax?