@tontonsb 2019-10-31
@donnysim which casts are we talking about?

I took a brief look at various things that might go on in a cast:

![image](https://user-images.githubusercontent.com/16481303/67958900-987c6700-fc00-11e9-90a9-f5cecc618815.png)

According to my previous results, the proposed solution would cost around 0.24-0.28 seconds per million directives parsed. That's four times slower than `(int) '12.12'` but 30% faster than `(string) 12.12` and 30 times faster than carbon instantiation that's also going on in casting. So one might say that having 30 visibility directives costs as much as casting `created_at` to date.