@tontonsb 2019-10-31
> @tontonsb how many other methods would have to be touched if we did the array structure syntax?

I am not sure this is a complete list, but I am aware of these:
- `getArrayableItems`  in `Illuminate/Database/Eloquent/Concerns/HasAttributes` should be tweaked a bit as it would have to use keys instead of values for some entries.
- `makeHidden` and `makeVisible` will have to be rewritten in .`Illuminate/Database/Eloquent/Concerns/HidesAttributes`.
- looks like I've advertised to remove `addHidden` in #30348 but forgot to include that change in PR, so that method should also either be rewritten or removed.