@donnysim 2019-10-31
It was half year ago or more, so don't remember the exact numbers, but we had a query `Task::...->pluck('id')` and it took around >1s  `toArray()` (there were A LOT of entries) because Task model had `$casts = ['id' => 'int']` (behind the scenes even pluck does model initiaion and casting if required). What we ended up doing was using `CONCAT` all ids then explode and array_map with intval with which we ended up around 100-200ms (can't remember the exact numbers, could've been sub 100ms too) just for retrieving ids as integers. We probably could've shaved more time with raw sql call.
XDebug profiler and CacheGrid were star players in debugging the cause.

This is kind of off topic so there's no reason to dig into it further as this has nothing to do with casting :)