@eli-s-r 2020-10-29
Yeah, this would be an awesome addition. Currently working on a project using lots of nested relationships where the fields to be hidden are highly variable from one page to another, and adding this feature would greatly reduce the amount of manual iteration necessary.