@OrkhanAlikhanov 2020-10-30
@eli-s-r I actually ended up overriding `makeVisible` method in the trait and using it in my base model.

```php
<?php

namespace App\Traits;

trait NestedVisible
{
    /**
     * Makes nested models visible. E.g makeVisible('resource.owner')
     * @param $attributes
     * @return $this
     */
    public function makeVisible($attributes)
    {
        $attributes = is_array($attributes) ? $attributes : func_get_args();

        foreach ($attributes as $attribute) {
            $nested = explode('.', $attribute, 2);

            if (count($nested) === 2) {
                $v = $nested[0];
                $this->$v->makeVisible($nested[1]);
            }
        }

        return parent::makeVisible($attributes);
    }
}
```

My base model.
```php
class Model extends EloquentModel
{
    use NestedVisible;
}
```