@Packoverflow 2022-08-23
Bump! At least an e.g. static Model::makeVisible([...]) would help a lot to set visible/hidden attributes of related Models at runtime :)

Love & Much appreciated.